[WIP] Port to using fs module for generic filesystem
Allows for reading from zip files for example.  This works now:

$ fontmake -u zip://NotoSansArmenian-Regular.zip

The reader is almost fully ported.  The writer needs more work.

This implementation uses the fs module directly.  In a finished
implementation we don't need to depend on that module.  We can
implement a very light fs.OSFS-equivalent, and leave it to user
to pass in fs objects if they prefer to do so.

On the other hand, relying on fs module is nice because of the
URL scheme (zip://, etc).  Or we can implement sniffing...

Fixes https://github.com/unified-font-object/ufoLib/issues/23